--- src/main/java/pro/trevor/tankgame/ui/IRpcHandler.java
@@ -1,8 +1,0 @@
-package pro.trevor.tankgame.ui;
-
-public interface IRpcHandler {
-    /**
-     * Indicate the rpc handler can accept more requests from the client
-     */
-    boolean canProcessRequests();
-}
